Make unexpected a full dependency.
Since the library was written npm has changed its behaviour around
installing peer dependencies to no longer implicitly pull them in.
The documentation does not mention having to install Unexpected,
instead mentioning just requiring expect-the-unexpected. It seems
that the intention was that this could be a drop in without extra
work, so follow through on this and make it self contained.