Realign with Unexpected.
Rename addAssertion() to addCustomAssertion() to avoid clash.
The library defined its own addAssertion() API which was not
compatible with Unexpected core. This meant an inability to
expose an interface that was compatible and opens the door
to future use of Unexpected plugins with the library.

The change has been made such that it includes clear messages
explaining the situation to existing users.